refactor: Extracting interface from Parser -> IFormParser;
