+++ EAutopay.Tests.Integration/FormTest.cs
@@ -9,13 +9,14 @@ namespace EAutopay.Tests.Integration
     [TestClass]
     public class FormTest
     {
-        IFormRepository _formRepository = new EAutopayFormRepository();
+        IFormRepository _formRepository;
 
         [TestInitialize]
         public void SetUp()
         {
             HttpContext.Current = Common.GetHttpContext();
             Common.Login();
+            _formRepository = new EAutopayFormRepository();
         }
 
         [TestCleanup]
@@ -25,6 +26,7 @@ namespace EAutopay.Tests.Integration
 
             Common.Logout();
             HttpContext.Current = null;
+            _formRepository = null;
         }
 
         [TestMethod]
