+++ EAutopay/Parsers/IFormParser.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+using EAutopay.Forms;
+
+namespace EAutopay.Parsers
+{
+    public interface IFormParser
+    {
+        List<Form> ExtractForms(string source);
+    }
+}
